fix: rename track to analytics
