+++ src/pages/_document.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 
-import { GtmNoScript } from '../utils/track';
+import { GtmNoScript } from '../utils/analytics';
 
 class MyDocument extends Document {
   static async getInitialProps(ctx) {
